Apply SSTA

+++ include/iq_network.h
@@ -20,7 +20,8 @@ public:
     void send_synapse();                    // proceed one timestep
     void printfile(FILE **fp);              // output potentials
     int set_biascurrent(int neuron_index, int biascurrent);
-    int set_neuron(int neuron_index, int rest, int threshold, int reset, int a, int b, int noise);
+    //int set_neuron(int neuron_index, int rest, int threshold, int reset, int a, int b, int noise);
+    int set_neuron(int neuron_index, int rest, int threshold, int reset, float a, float b, int noise);
     int set_weight(int pre, int post, int weight, int tau);
     int potential(int neuron_index);
     int spike_count(int neuron_index);
